semantic_matching/requirements.txt
sentence-transformers
scikit-learn
numpy
python-dotenv
hf_transfer
